Add files via upload
Under maintenance mode, the user could control the servos manually as well as receive data from the distance sensor, colour sensor or card reader in the text box. The form uses a combo box to select from available COM ports. Once opened, the form is fully operational with the MBED device. There are three positions options available to test for each servo; the face angle can be set to 0°, 90° or 180° and the prize motor can be set up, down or up then down.